Check for ack before taking measurement
This solves the second problem. By moving the acknowledgement check and
subsequent updating of the list, the receiver now has 60 seconds to get
the message and send an ack. While it is still theoretically possible
that it doesn't manage this in time because of network issues, this has
become extremely unlikely. If no ack has been received in 60 seconds,
the connection has probably broken and there is no ack coming at all.

As such, it should no longer be possible to accumulate acks from
previous, delayed transmissions in the message queue and every received
ack concerns the current contents of the measurement queue.